# Conflicts:
#	cache/pom.xml
#	pom.xml
#	runelite-api/pom.xml
#	runelite-client/pom.xml
#	runelite-client/src/main/java/net/runelite/client/plugins/gpu/GpuIntBuffer.java
#	runelite-client/src/main/java/net/runelite/client/plugins/gpu/GpuPlugin.java
#	runelite-jshell/pom.xml
#	runelite-maven-plugin/pom.xml

+++ runelite-api/src/main/java/net/runelite/api/Mesh.java
@@ -48,9 +48,9 @@ public interface Mesh<T extends Mesh<T>>
 	List<Triangle> getTriangles();
 
 	int getVerticesCount();
-	int[] getVerticesX();
-	int[] getVerticesY();
-	int[] getVerticesZ();
+	float[] getVerticesX();
+	float[] getVerticesY();
+	float[] getVerticesZ();
 
 	int getFaceCount();
 	int[] getFaceIndices1();
